requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cmake==3.18.4.post1
cycler==0.10.0
dataclasses==0.6
dl==0.1.0
dlib==19.22.0
Flask==1.1.2
future==0.18.2
idna==2.10
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
list-print==1.0.1
MarkupSafe==1.1.1
matplotlib==3.4.1
numpy==1.20.2
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
scipy==1.6.2
six==1.15.0
to==0.3
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
